Create Getin 1.0
Get.in: Virtual Reality Ticketing System for Events

Get.in is an innovative online ticketing platform built on the world chain that offers virtual reality attendance for various events, including sports games, concerts, gaming tournaments, and more. It aims to provide an immersive experience for individuals who are unable to physically attend an event but still want to be part of the excitement.

With Get.in, users can quickly purchase tickets for listed events using their world ID and pay with worldcoin, a digital currency. By logging in and making payment through their world ID, users gain anonymous entry and privacy throughout the event, all from the comfort of their own homes using VR glasses.

To ensure a high-quality experience, events listed on Get.in require VR broadcasting. This involves using special cameras capable of capturing events in 3D, 4D, or even 5D, providing attendees with a closer and clearer view of the action. This benefits individuals who prefer a closer view of the stage or have visual impairments, enhancing their overall event experience.

To keep the platform active, Get.in can also offer free events, such as replays, wildlife content, and other agreed-upon content. These free events can be accessed by anyone with a world ID and a pair of VR glasses, allowing users to enjoy the feeling of attending a real-time event even if they haven't purchased a ticket.

To support the platform, partnerships can be established with VR glasses manufacturers, broadcasters, and reliable internet service providers. Additionally, a 4% taxation on event organizers can be implemented to generate revenue for the project.